requirements.txt
Flask==2.0.2
Flask-Bcrypt==0.7.1
Flask-JWT-Extended==4.3.1
Flask-Mail==0.9.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0